# openapi.yml
openapi: 3.1.0
info:
  title: Camplx Campus Circular Marketplace API
  version: 1.0.0
  description: 'Canonical backend contract for CAMPLX (CampusSwap AI). Facilitates
    peer-to-peer campus reuse, multi-angle computer vision product inspection, evidence-backed
    appraisal pricing, no-payment reservations, campus want-board ("Need It"), e-waste
    collection optimization, and verified sustainability metrics.

    '
servers:
- url: http://localhost:8000/api/v1
  description: Local FastAPI Development Server (iOS Simulator / Web)
- url: http://10.0.2.2:8000/api/v1
  description: Android Emulator Gateway
- url: https://api.camplx.internal/api/v1
  description: Staging / Production Backend Gateway
...
paths:
...
  /listings:
    get:
      tags:
      - Listings
      operationId: listMarketplaceListings
      summary: Browse and search campus listings
      description: Returns paginated active marketplace listings filtered by campus,
        category, keyword, and condition.
      parameters:
      - name: campus_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: category
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ItemCategoryEnum'
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ListingStatusEnum'
        description: Defaults to 'active' if unspecified.
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: listing_type
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ListingTypeEnum'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          minimum: 1
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 20
          maximum: 100
      responses:
        '200':
          description: Paginated listings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListingsResponse'
        '400':
          description: Bad filter parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Listings
      operationId: createMarketplaceListing
      summary: Publish a new listing
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Listing created and published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetailResponse'
        '400':
          description: Invalid listing payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /listings/{listing_id}:
    parameters:
    - name: listing_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Listings
      operationId: getListingDetail
      summary: Retrieve complete details for a listing
      description: Returns listing attributes, images, linked AI inspection findings,
        and active reservation state.
      responses:
        '200':
          description: Listing details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetailResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Listings
      operationId: updateMarketplaceListing
      summary: Edit own listing attributes (Restricted Status Modification)
      description: 'Allows updating title, description, asking price, and toggling
        between active and inactive. Listing edits strictly cannot transition items
        to ''reserved'', ''sold'', or ''donated''; those transitions are exclusively
        driven by atomic reservation and completion workflows.

        '
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListingRequest'
      responses:
        '200':
          description: Listing updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetailResponse'
        '400':
          description: Invalid update or attempt to bypass reservation workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Not the listing owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Listings
      operationId: deleteMarketplaceListing
      summary: Remove a listing
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '204':
          description: Listing deleted
        '403':
          description: Forbidden (Not the listing owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sellers/{seller_id}/listings:
    parameters:
    - name: seller_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Listings
      operationId: getSellerPublicProfileAndListings
      summary: Retrieve public seller profile and active inventory
      description: Returns public-safe seller details (excluding private institutional
        email) and their active inventory.
      responses:
        '200':
          description: Seller public profile and active inventory retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerProfileWithListingsResponse'
        '404':
          description: Seller not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /need-requests:
    get:
      tags:
      - Need It Requests
      operationId: listNeedItRequests
      summary: Browse open campus want-requests
      parameters:
      - name: category
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ItemCategoryEnum'
      - name: urgency
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/NeedUrgencyEnum'
      responses:
        '200':
          description: List of campus need requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NeedRequestRecord'
    post:
      tags:
      - Need It Requests
      operationId: createNeedItRequest
      summary: Post a request for a needed item
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNeedRequestPayload'
      responses:
        '201':
          description: Need request published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeedRequestRecord'
  /need-requests/{request_id}:
    parameters:
    - name: request_id
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - Need It Requests
      operationId: deleteNeedItRequest
      summary: Delete or close a personal need request
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      responses:
        '204':
          description: Request deleted
        '403':
          description: Forbidden (Not the requester)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /need-requests/{request_id}/matches:
    parameters:
    - name: request_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - Need It Requests
      operationId: getMatchesForNeedRequest
      summary: Query active campus marketplace listings matching this request
      description: Returns heuristic keyword and category matches found in current
        marketplace inventory.
      responses:
        '200':
          description: Matched catalog items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NeedMatchResult'
...
  /e-waste/requests:
    get:
      tags:
      - E-Waste Optimization
      operationId: listEWasteRequests
      summary: List all campus e-waste collection requests
      parameters:
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/EWasteLifecycleStatusEnum'
      - name: device_category
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/EWasteDeviceCategoryEnum'
      responses:
        '200':
          description: List of logged collection requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EWasteRequestRecord'
    post:
      tags:
      - E-Waste Optimization
      operationId: submitEWasteRequest
      summary: Submit end-of-life or repairable hardware for campus recovery
      description: 'Submits physical electronics (laptops, phones, power banks, cables,
        etc.) specifying physical device category, fault description, quantity, campus
        pickup coordinates, and preferred time slot.

        '
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEWasteSubmissionRequest'
      responses:
        '201':
          description: E-waste pickup logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EWasteRequestRecord'
  /e-waste/cluster-and-optimize:
    post:
      tags:
      - E-Waste Optimization
      operationId: optimizeEWasteCollectionRoutes
      summary: Run K-Means spatial clustering and TSP route sequencing
      description: 'Partitions pending coordinates into zones via Scikit-Learn K-Means,
        computes shortest pickup manifests via Haversine TSP nearest-neighbor algorithm,
        and updates status to ''scheduled''.

        '
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      parameters:
      - name: zones_needed
        in: query
        required: false
        schema:
          type: integer
          default: 2
          minimum: 1
          maximum: 10
      responses:
        '200':
          description: Optimized collection routes generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EWasteOptimizationPlanResponse'
        '400':
          description: Insufficient coordinates to compute clustering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema: *id001
  /e-waste/{request_id}/lifecycle:
    parameters:
    - name: request_id
      in: path
      required: true
      schema:
        type: string
    patch:
      tags:
      - E-Waste Optimization
      operationId: updateEWasteLifecycleStatus
      summary: Advance request lifecycle stage
      description: 'Tracks progression: pending -> scheduled -> collected -> handed_over.
        Terminal handovers record certified downstream e-waste recycler partner IDs.

        '
      security:
      - BearerAuth: []
      - DemoUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEWasteStatusRequest'
      responses:
        '200':
          description: Lifecycle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EWasteRequestRecord'
        '400':
          description: Invalid lifecycle transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: PocketBase or standard JWT token.
    DemoUserHeader:
      type: apiKey
      in: header
      name: X-Camplx-Demo-User-Id
      description: 'Development and demo identity bypass header (e.g. a0000000-0000-0000-0000-000000000001).
        Strictly disallowed and ignored in production environments.

        '
  schemas:
    ItemCategoryEnum:
      type: string
      enum:
      - electronics
      - books
      - cycles
      - furniture
      - hostel_essentials
      - other
    ListingTypeEnum:
      type: string
      enum:
      - sell
      - donate
    ListingStatusEnum:
      type: string
      enum:
      - active
      - reserved
      - sold
      - donated
      - inactive
...
    ConditionGradeEnum:
      type: string
      enum:
      - Like New
      - Good
      - Fair
      - Needs Repair
    LifecycleCategoryEnum:
      type: string
      enum:
      - Reusable
      - Repairable
      - End-of-life
    EWasteDeviceCategoryEnum:
      type: string
      enum:
      - laptops_and_computers
      - phones_and_tablets
      - batteries_and_power_banks
      - chargers_and_cables
      - printers_and_peripherals
      - audio_and_accessories
      - other_electronics
...
    NeedUrgencyEnum:
      type: string
      enum:
      - low
      - normal
      - urgent
    EWasteLifecycleStatusEnum:
      type: string
      enum:
      - pending
      - scheduled
      - collected
      - handed_over
...
    AttributeProvenanceEnum:
      type: string
      enum:
      - visually_verified
      - seller_supplied
      - catalog_inferred
      - ai_estimated
      description: Provenance classification for product attributes, distinguishing
        observations from inference.
    ErrorResponse:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          example: INVALID_STATE_TRANSITION
        message:
          type: string
          example: This listing is already reserved by another student.
        details:
          type: object
          additionalProperties: true
...
    InspectionImageRecord:
      type: object
      required:
      - id
      - image_url
      - display_order
      - angle_label
      properties:
        id:
          type: string
          example: img_capture_01
        image_url:
          type: string
          format: uri
          example: https://api.camplx.internal/media/inspections/raw_01.jpg
        annotated_image_url:
          type:
          - string
          - 'null'
          format: uri
          description: URL of derived image asset with overlay bounding boxes rendered
            for this specific capture angle.
          example: https://api.camplx.internal/media/inspections/annotated_01.jpg
        display_order:
          type: integer
          example: 0
        angle_label:
          type: string
          example: front
...
    CreateListingRequest:
      type: object
      required:
      - title
      - category
      - condition
      - listing_type
      - image_urls
      properties:
        title:
          type: string
          example: Casio FX-991EX ClassWiz Scientific Calculator
        brand:
          type:
          - string
          - 'null'
          example: Casio
        model:
          type:
          - string
          - 'null'
          example: FX-991EX
        description:
          type:
          - string
          - 'null'
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        condition:
          $ref: '#/components/schemas/ConditionGradeEnum'
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        asking_price:
          type:
          - number
          - 'null'
          description: Required when listing_type is 'sell'. Must be null for 'donate'.
          example: 950
        inspection_id:
          type:
          - string
          - 'null'
        technical_specifications:
          type: object
          additionalProperties:
            type: string
          example:
            Power: Solar & Battery
            Functions: '552'
        specification_provenance:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AttributeProvenanceEnum'
        image_urls:
          type: array
          items:
            type: string
          minItems: 1
          example:
          - https://api.camplx.internal/media/inspections/raw_01.jpg
    UpdateListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type:
          - string
          - 'null'
        asking_price:
          type:
          - number
          - 'null'
        status:
          $ref: '#/components/schemas/SellerListingStatusUpdateEnum'
    ListingSummary:
      type: object
      required:
      - id
      - seller_id
      - title
      - category
      - condition
      - listing_type
      - status
      - primary_image_url
      - created_at
      properties:
        id:
          type: string
        seller_id:
          type: string
        title:
          type: string
        brand:
          type:
          - string
          - 'null'
          example: Casio
        model:
          type:
          - string
          - 'null'
          example: FX-991EX
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        condition:
          $ref: '#/components/schemas/ConditionGradeEnum'
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        asking_price:
          type:
          - number
          - 'null'
        status:
          $ref: '#/components/schemas/ListingStatusEnum'
        carbon_savings_g:
          type: integer
          example: 3500
        primary_image_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
    ListingDetailResponse:
      type: object
      required:
      - id
      - seller_id
      - seller_name
      - campus_id
      - title
      - category
      - condition
      - listing_type
      - status
      - images
      - created_at
      properties:
        id:
          type: string
        seller_id:
          type: string
        seller_name:
          type: string
          example: Aarav Sharma
        campus_id:
          type: string
        title:
          type: string
        brand:
          type:
          - string
          - 'null'
          example: Casio
        model:
          type:
          - string
          - 'null'
          example: FX-991EX
        description:
          type:
          - string
          - 'null'
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        condition:
          $ref: '#/components/schemas/ConditionGradeEnum'
        listing_type:
          $ref: '#/components/schemas/ListingTypeEnum'
        asking_price:
          type:
          - number
          - 'null'
        status:
          $ref: '#/components/schemas/ListingStatusEnum'
        carbon_savings_g:
          type: integer
          example: 3500
        technical_specifications:
          type: object
          additionalProperties:
            type: string
          example:
            Power: Solar + Battery
        specification_provenance:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AttributeProvenanceEnum'
        images:
          type: array
          items:
            $ref: '#/components/schemas/InspectionImageRecord'
        inspection:
          anyOf:
          - $ref: '#/components/schemas/InspectionAnalysisResultResponse'
          - type: 'null'
        pricing_intelligence:
          anyOf:
          - $ref: '#/components/schemas/PricingIntelligence'
          - type: 'null'
        active_reservation:
          anyOf:
          - $ref: '#/components/schemas/ReservationRecord'
          - type: 'null'
        created_at:
          type: string
          format: date-time
    PaginatedListingsResponse:
      type: object
      required:
      - items
      - total
      - page
      - per_page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ListingSummary'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    SellerProfileWithListingsResponse:
      type: object
      required:
      - seller
      - listings
      properties:
        seller:
          $ref: '#/components/schemas/PublicSellerProfileResponse'
        listings:
          type: array
          items:
            $ref: '#/components/schemas/ListingSummary'
...
    CreateNeedRequestPayload:
      type: object
      required:
      - title
      - category
      properties:
        title:
          type: string
          example: Casio 991EX for semester end exams
        note:
          type:
          - string
          - 'null'
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        urgency:
          $ref: '#/components/schemas/NeedUrgencyEnum'
          default: normal
    NeedRequestRecord:
      type: object
      required:
      - id
      - requester_id
      - requester_name
      - title
      - category
      - urgency
      - created_at
      properties:
        id:
          type: string
        requester_id:
          type: string
        requester_name:
          type: string
        title:
          type: string
        note:
          type:
          - string
          - 'null'
        category:
          $ref: '#/components/schemas/ItemCategoryEnum'
        urgency:
          $ref: '#/components/schemas/NeedUrgencyEnum'
        created_at:
          type: string
          format: date-time
    NeedMatchResult:
      type: object
      required:
      - listing
      - match_reason
      properties:
        listing:
          $ref: '#/components/schemas/ListingSummary'
        match_reason:
          type: string
          example: Keyword 'casio' matched title in electronics category.
    CreateEWasteSubmissionRequest:
      type: object
      required:
      - description
      - device_category
      - quantity
      - latitude
      - longitude
      - location_name
      - preferred_slot
      properties:
        description:
          type: string
          example: Dell Inspiron with dead motherboard & swollen battery
        device_category:
          $ref: '#/components/schemas/EWasteDeviceCategoryEnum'
        lifecycle_assessment:
          $ref: '#/components/schemas/LifecycleCategoryEnum'
          default: End-of-life
        quantity:
          type: integer
          minimum: 1
          example: 1
        latitude:
          type: number
          format: double
          example: 18.6251
        longitude:
          type: number
          format: double
          example: 73.8198
        location_name:
          type: string
          example: Tech Block B, DPU Campus
        preferred_slot:
          type: string
          example: Morning (10:00 AM - 1:00 PM)
    EWasteRequestRecord:
      type: object
      required:
      - id
      - student_id
      - description
      - device_category
      - lifecycle_assessment
      - quantity
      - location_name
      - status
      - estimated_carbon_kg
      - created_at
      properties:
        id:
          type: string
        student_id:
          type: string
        description:
          type: string
        device_category:
          $ref: '#/components/schemas/EWasteDeviceCategoryEnum'
        lifecycle_assessment:
          $ref: '#/components/schemas/LifecycleCategoryEnum'
        quantity:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        location_name:
          type: string
        preferred_slot:
          type: string
        status:
          $ref: '#/components/schemas/EWasteLifecycleStatusEnum'
        zone_cluster_id:
          type:
          - integer
          - 'null'
        pickup_sequence_order:
          type:
          - integer
          - 'null'
        estimated_carbon_kg:
          type: number
          example: 34.5
        created_at:
          type: string
          format: date-time
    UpdateEWasteStatusRequest:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/EWasteLifecycleStatusEnum'
        handler_notes:
          type:
          - string
          - 'null'
        certified_partner_id:
          type:
          - string
          - 'null'
    EWasteOptimizationPlanResponse:
      type: object
      required:
      - message
      - total_stops_optimized
      - routes_generated
      properties:
        message:
          type: string
          example: Clustering complete and routes sequenced.
        total_stops_optimized:
          type: integer
          example: 4
        routes_generated:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
            description: Ordered sequence of request IDs.
          example:
            Zone 0:
            - ewaste-1
            - ewaste-2
            Zone 1:
            - ewaste-3
...